pyproject.toml
[project]
name = "open-rag"
version = "0.1.0"
description = "Open-source RAG system using Hugging Face LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "openai>=1.0.0",
    "pdfplumber>=0.10.0",
]
